+++ CHANGELOG.md
@@ -14,6 +14,12 @@ when methods returned RecordList-type objects.
 to historical Whois (ownership) records. Includes more consistent implementation of
 RecordList functionality and better pandas dataframe support for both historical Whois and 
 field-level Whois searches. 
+- New `impacted_attack_surfaces` property of vulnerability articles (`VulnArticle`) filters 
+the list of third-party vendors to only those with at least one observation. The Illuminate 
+API returns all attack surfaces associated with an API key regardless of whether they are 
+impacted; the complete list is still available in the `attack_surfaces` property. Also updated
+the `info` view of the Pandas dataframe on a vulnerability article so the `impacts` column
+shows the count of impacted attack surfaces.
 
 
 #### Bug Fixes
@@ -24,6 +30,9 @@ field-level Whois searches.
 record list were only counting high-priority insights. 
 - Fixed issue that caused an exception when trying to generate a dictionary view of an
 AttackSurfaceComponent (detection). 
+- Removed reference to non-existant field in `VulnArticle` that was causing an exception when
+rendering a vulnerability article as a dictionary with the `as_dict` property.
+- Handle vuln articles with no impacted assets without raising an exception.